requirements.txt
asgiref==3.5.0
attrs==21.4.0
Django==3.2
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
django-simple-history==3.0.0
iniconfig==1.1.1
mysqlclient==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
